pyproject.toml
[tool.poetry]
name = "billing-system"
version = "0.1.0"
description = ""
authors = ["Shivam Gupta <shivam.gupta02210@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
black = "^24.8.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
django-celery-beat = "^2.7.0"
celery = "^5.4.0"
redis = "^5.0.8"
xhtml2pdf = "^0.2.16"
drf-spectacular = "^0.27.2"
faker = "^28.4.1"
faker-commerce = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
